[2.x] feat: Support doc / skip := true
**Problem**
Setting `doc / skip := true` had no effect because the doc task
did not check the skip setting. Only compile, update, publish, and
publishLocal respected it.

**Solution**
Move the doc task implementation to Compiler.scala and use Def.taskIf
to properly short-circuit evaluation when skip is true. When skipped,
the compilers, sources, and other heavy dependencies are not evaluated.
Update the skip key description to include doc.

Fixes #5525
