+++ sbt-app/src/sbt-test/project/skip-doc/build.sbt
@@ -1,0 +1,5 @@
+lazy val root = (project in file("."))
+  .settings(
+    scalaVersion := "3.7.4",
+    doc / skip := true,
+  )
