+++ sbt-app/src/sbt-test/project/skip-doc/src/main/scala/A.scala
@@ -1,0 +1,2 @@
+object A:
+  def hello: String = "hello"
